# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agent Browser API",
        "description": "Web browsing and search capabilities powered by agent-browser",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/search": {
            "post": {
                "operationId": "searchWeb",
                "summary": "Search the web",
                "description": "Search the web using Google and return structured results",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "query"
                                ],
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "description": "The search query"
                                    },
                                    "maxResults": {
                                        "type": "integer",
                                        "description": "Maximum number of results to return",
                                        "default": 5,
                                        "minimum": 1,
                                        "maximum": 20
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "query": {
                                            "type": "string"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "link": {
                                                        "type": "string"
                                                    },
                                                    "snippet": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/browse": {
            "post": {
                "operationId": "browseUrl",
                "summary": "Browse and extract content from a URL",
                "description": "Fetch a webpage and extract specific content using CSS selectors",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "uri",
                                        "description": "The URL to browse"
                                    },
                                    "selector": {
                                        "type": "string",
                                        "description": "CSS selector to target specific elements",
                                        "default": "body"
                                    },
                                    "extract": {
                                        "type": "string",
                                        "description": "What to extract (e.g., 'text', 'html', or custom like 'title:h1, content:p')",
                                        "default": "text"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Extracted content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "url": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "object",
                                            "description": "Extracted content based on selector and extract parameters"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/screenshot": {
            "post": {
                "operationId": "getScreenshot",
                "summary": "Take a screenshot of a webpage",
                "description": "Capture a screenshot of the specified URL and return it as base64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "uri",
                                        "description": "The URL to screenshot"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Screenshot data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "url": {
                                            "type": "string"
                                        },
                                        "screenshot": {
                                            "type": "string",
                                            "description": "Base64-encoded screenshot image"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}